multipath-tools, nixos/multipath: add support for stage-1 boot
By default multipath-tools is compiled to require systemd which prevents
it from being used in stage 1 boot. This patch adds a new `-1` flag to
`multipathd` which disables all interactions with systemd and adds the
necessary logic to copy multipath configuration to the initrd and launch
`multipathd -1` in stage 1.

This patch makes it possible to bootstrap the system and mount a multipath
device over /nix.